+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v7.2.0
 * **[2021-12-27 12:05:54 CDT]** Use the official PHP 8.1 builds now.
 * **[2021-12-27 12:06:15 CDT]** Include ext-imap and ext-ssh2.
+* **[2021-12-27 17:18:50 CDT]** (#10) Added support for the Ioncube Decoder.
 
 ## v7.1.2
 * **[2021-12-24 23:27:47 CDT]** Added support for composer v2.2.0+.